Conflicts:
	addons/skin.confluence/720p/ViewsFileMode.xml
	addons/skin.confluence/720p/ViewsVideoLibrary.xml
	addons/skin.touched
	language/English/strings.po
	language/Russian/strings.po
	project/VS2010Express/XBMC for Windows.sln
	project/VS2010Express/XBMC.vcxproj
	project/VS2010Express/XBMC.vcxproj.filters
	xbmc/Application.cpp
	xbmc/GUIInfoManager.cpp
	xbmc/GUIInfoManager.h
	xbmc/cores/VideoRenderers/LinuxRendererGL.cpp
	xbmc/cores/VideoRenderers/LinuxRendererGL.h
	xbmc/cores/VideoRenderers/RenderManager.cpp
	xbmc/cores/VideoRenderers/RenderManager.h
	xbmc/cores/dvdplayer/DVDClock.h
	xbmc/cores/dvdplayer/DVDDemuxers/DVDDemux.h
	xbmc/filesystem/HDDirectory.cpp
	xbmc/filesystem/windows/WINSMBDirectory.cpp
	xbmc/osx/atv2/XBMCAppliance.mm
	xbmc/osx/atv2/substrate.h
	xbmc/settings/GUISettings.cpp
	xbmc/settings/GUISettings.h
	xbmc/settings/VideoSettings.h
	xbmc/settings/windows/GUIWindowSettingsCategory.cpp
	xbmc/settings/windows/GUIWindowSettingsCategory.h
	xbmc/utils/CharsetConverter.cpp
	xbmc/utils/CharsetConverter.h
	xbmc/utils/DatabaseUtils.cpp
	xbmc/utils/SaveFileStateJob.h
	xbmc/utils/StdString.h
	xbmc/utils/StreamDetails.h
	xbmc/video/VideoDatabase.cpp
	xbmc/video/VideoDatabase.h
	xbmc/video/dialogs/GUIDialogAudioSubtitleSettings.cpp
	xbmc/video/dialogs/GUIDialogVideoSettings.cpp
	xbmc/video/windows/GUIWindowVideoBase.cpp
	xbmc/windowing/windows/WinEventsWin32.cpp
	xbmc/windows/GUIMediaWindow.cpp

+++ xbmc/cores/dvdplayer/DVDMessage.cpp
@@ -1,6 +1,6 @@
 /*
  *      Copyright (C) 2005-2013 Team XBMC
- *      http://www.xbmc.org
+ *      http://xbmc.org
  *
  *  This Program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License as published by
@@ -57,7 +57,7 @@ CDVDMsgGeneralSynchronize::~CDVDMsgGeneralSynchronize()
   delete m_p;
 }
 
-bool CDVDMsgGeneralSynchronize::Wait(unsigned long milliseconds, unsigned int source)
+bool CDVDMsgGeneralSynchronize::Wait(unsigned int milliseconds, unsigned int source)
 {
   if(source == 0)
     source = SYNCSOURCE_OWNER;
